ADD services with response the nearest cinema

+++ FilmsAPI.Dto/Dtos/CinemaDto.cs
@@ -1,10 +1,8 @@
-using System.ComponentModel.DataAnnotations;
-
 namespace FilmsAPI.Dto
 {
     public class CinemaDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
